user login and registration functionality implemented

+++ base/views.py
@@ -1,8 +1,10 @@
 from tkinter.tix import Form
+from django.http import HttpResponse
 from django.shortcuts import redirect, render
 from django.db.models import Q
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from .models import Room, Topic
 from .forms import RoomForm
@@ -28,7 +30,7 @@ def home(request):
     topics = Topic.objects.all()
     room_count = rooms.count()
     print(rooms)
-    context = {'rooms': rooms, 'topics': topics, "room_count": room_count }
+    context = {'rooms': rooms, 'topics': topics, "room_count": room_count}
     return render(request, 'base/home.html', context)
 
 
@@ -43,6 +45,7 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -57,9 +60,14 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here.")
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -71,12 +79,20 @@ def updateRoom(request, pk):
 
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here.")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj': room})
 
+
 def loginPage(request):
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect("home")
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -84,17 +100,22 @@ def loginPage(request):
             user = User.objects.get(username=username)
         except:
             messages.error(request, 'Username or password does not exists.')
-        user = authenticate(request, username=username,password=password)
+        user = authenticate(request, username=username, password=password)
 
         if user is not None:
-            login(request, user) # add session in db and in browser
+            login(request, user)  # add session in db and in browser
             return redirect('home')
         else:
             messages.error(request, 'Username or password does not exists.')
-   
-    context = {}
+
+    context = {"page":page}
     return render(request, 'base/login_register.html', context)
+
+def registerPage(request):
+    page = "register"
+    context = {"page":page}
+    return render(request,'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
-    return redirect("home")    
+    return redirect("home")
